changefeedccl: move cloudStorageSink into its own file
Pure code movement, no changes.

Release note: None
